[OF-1375] refac: updates tests, docs, names
