+++ Library/include/CSP/Multiplayer/SpaceEntitySystem.h
@@ -353,7 +353,7 @@ public:
 
 	/// @brief Gets all hierarchies for a space
 	/// @param Callback SequenceHierarchiesResultCallback : A callback when the asynchronous task finishes.
-	CSP_ASYNC_RESULT void GetAllSequenceHierarchies(SequenceHierarchiesResultCallback Callback);
+	CSP_ASYNC_RESULT void GetAllSequenceHierarchies(SequenceHierarchyCollectionResultCallback Callback);
 
 	/// @brief Deletes an entity hierarchy for a given parent entity id. Pass null to delete the root hierarchy.
 	/// @param ParentId Optional<uint64_t> : An optional parent. Pass null to get the root hierarchy.
@@ -454,8 +454,6 @@ private:
 	bool IsInitialised = false;
 
 	SequenceHierarchyChangedCallbackHandler SequenceHierarchyChangedCallback;
-
-	systems::SequenceSystem* SequenceSystem;
 };
 
 } // namespace csp::multiplayer
