+++ Library/include/CSP/Systems/Sequence/Sequence.h
@@ -20,8 +20,10 @@
 namespace csp::systems
 {
 
+CSP_START_IGNORE
 // Prefix needed when storing multiplayer unsigned integer ids in keys
 static inline const csp::common::String SequenceIdPrefix = "m_Id_";
+CSP_END_IGNORE
 
 /// @ingroup Sequence System
 /// @brief A basic class abstraction for a sequence, including key, and reference variables, and items.
