+++ Library/src/Multiplayer/EventSerialisation.h
@@ -128,6 +128,10 @@ private:
 	SequenceChangedParams EventParams;
 };
 
+/// A deserialiser for getting SequenceHierarchy data from an event:
+/// UpdateType - The update type for the Sequence Hierarchy: Created, Updated, Deleted
+/// ParentId - The parent id of the Sequence
+/// IsRoot - Whether this is the root hierarchy of the space
 class SequenceHierarchyChangedEventDeserialiser : public EventDeserialiser
 {
 public:
